requirements.txt
streamlit
openai
Ipython
langgraph
langchain_openai
python-dotenv